02_Backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "FastAPI backend for Unified Portfolio Agent"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "openai-agents>=0.0.1",
    "python-dotenv>=1.0.0",
    "python-decouple>=3.8",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "openai-chatkit>=0.0.1",
    "rich>=13.0.0",
    "google-generativeai>=0.8.0",
    "google-auth>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
